# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API documentation for the zODC Masterflow Service. It provides endpoints for managing workflows, requests, forms, and nodes.",
        "title": "zODC Masterflow Service API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1  // Adjusted BasePath for clarity (e.g., /api/v1) - Ensure routes match this prefix",
    "paths": {
        "/services/forms/data/{formDataId}": {
            "get": {
                "description": "Retrieves detailed data of a specific submitted form by its ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "FormData"
                ],
                "summary": "Get form data by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Form Data ID",
                        "name": "formDataId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.JiraFormDetailResponse"
                        }
                    },
                    "400": {
                        "description": "error: Invalid ID or service error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/services/forms/templates": {
            "get": {
                "description": "Retrieves a list of form templates, optionally filtered by category ID and search query.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Forms"
                ],
                "summary": "Find all form templates",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by category ID",
                        "name": "categoryId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search query for form template names or descriptions",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/responses.FormTemplateFindAll"
                            }
                        }
                    },
                    "500": {
                        "description": "Error message for internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Adds a new form template to the system based on the provided data.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Forms"
                ],
                "summary": "Create a new form template",
                "parameters": [
                    {
                        "description": "Form Template Creation Request",
                        "name": "formTemplate",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.FormTemplateCreate"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "message: Form created successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Error message for bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Error message for internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/services/forms/templates/edit-profile": {
            "get": {
                "description": "Retrieves the form template used for editing a user's profile.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Forms"
                ],
                "summary": "Get edit profile form template",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.FormTemplateDetails"
                        }
                    },
                    "400": {
                        "description": "error: Service error or unable to retrieve the form template",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/services/forms/templates/performance-evaluate": {
            "get": {
                "description": "Retrieves the form template used for performance evaluations.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Forms"
                ],
                "summary": "Get performance evaluation form template",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.FormTemplateDetails"
                        }
                    },
                    "400": {
                        "description": "error: Service error or unable to retrieve the form template",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/services/forms/templates/{formTemplateId}": {
            "put": {
                "description": "Updates the details of a specific form template by its ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Forms"
                ],
                "summary": "Update a form template",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Form Template ID",
                        "name": "formTemplateId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Form Template update payload",
                        "name": "formTemplate",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.FormTemplateUpdate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "message: Form template updated successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "error: Invalid ID, bad request, or service error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/services/forms/templates/{formTemplateId}/details": {
            "get": {
                "description": "Retrieves detailed information of a specific form template using its ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Forms"
                ],
                "summary": "Get form template detail by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Form Template ID",
                        "name": "formTemplateId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.FormTemplateDetails"
                        }
                    },
                    "400": {
                        "description": "error: Invalid formTemplateId or service error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/services/services/categories": {
            "get": {
                "description": "Trả về danh sách tất cả các category có trong hệ thống.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Categories"
                ],
                "summary": "Lấy danh sách tất cả Category",
                "responses": {
                    "200": {
                        "description": "Danh sách Category",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/responses.CategoryFindAll"
                            }
                        }
                    },
                    "500": {
                        "description": "Lỗi Server Nội Bộ",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "requests.FormTemplateCreate": {
            "type": "object"
        },
        "requests.FormTemplateUpdate": {
            "type": "object"
        },
...
        "responses.CategoryFindAll": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "isActive": {
                    "type": "boolean"
                },
                "key": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "responses.FormTemplateDetails": {
            "type": "object",
            "properties": {
                "fields": {
                    "type": "array",
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/definitions/responses.FormTemplateFieldsFindAll"
                        }
                    }
                },
                "template": {
                    "$ref": "#/definitions/responses.FormTemplateFindAll"
                }
            }
        },
        "responses.FormTemplateFieldsFindAll": {
            "type": "object",
            "properties": {
                "advancedOptions": {
                    "type": "object",
                    "additionalProperties": true
                },
                "category": {
                    "type": "string"
                },
                "colNum": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "fieldId": {
                    "type": "string"
                },
                "fieldName": {
                    "type": "string"
                },
                "fieldType": {
                    "type": "string"
                },
                "formId": {
                    "type": "integer"
                },
                "icon": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "required": {
                    "type": "boolean"
                },
                "title": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "responses.FormTemplateFindAll": {
            "type": "object",
            "properties": {
                "category": {
                    "$ref": "#/definitions/responses.CategoryFindAll"
                },
                "createdAt": {
                    "type": "string"
                },
                "dataSheet": {
                    "type": "object",
                    "additionalProperties": true
                },
                "decoration": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "fileName": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "templateId": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "version": {
                    "type": "integer"
                },
                "versionId": {
                    "type": "integer"
                }
            }
        },
...
        "responses.JiraFormDetailResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/responses.NodeFormDataResponse"
                    }
                },
                "fields": {
                    "type": "array",
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/definitions/responses.FormTemplateFieldsFindAll"
                        }
                    }
                },
                "template": {
                    "$ref": "#/definitions/responses.FormTemplateFindAll"
                }
            }
        },
...
        "responses.NodeFormDataResponse": {
            "type": "object",
            "properties": {
                "fieldId": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}